# swagger.yaml
openapi: 3.0.0
info:
  title: obs_api
  contact: {}
  version: "1.0"
servers:
  - url: http://localhost:5001/api/v1
    variables: {}
paths:
  /jobs:
    get:
      tags:
        - Jobs
      summary: get jobs created by user
      operationId: getjobscreatedbyuser
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
    post:
      tags:
        - Jobs
      summary: create job
      operationId: createjob
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/createjobrequest"
                - example:
                    company: MLSWorldwide
                    position: Junior Next.js developer
            example:
              company: MLSWorldwide
              position: Junior Next.js developer
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []
...
components:
  schemas:
...
    createjobrequest:
      title: createjobrequest
      required:
        - company
        - position
      type: object
      properties:
        company:
          type: string
        position:
          type: string
      example:
        company: MLSWorldwide
        position: Junior Next.js developer
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...